Refine local dev: "daemonize" port forwarding
Avoid having mixed up output in the shell which launched the dev
cluster, and keep those jobs from being killed by a HUP from that
shell. To do so, put the port forwarding commands in a small script
which runs them backgrounded and disowns.

This sends the output for the port forwarding to log files, so that
it can be inspected.